+++ src/main/java/be/ceau/chart/data/PieDataset.java
@@ -19,8 +19,8 @@ import java.math.BigDecimal;
 import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility;
+import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 
 import be.ceau.chart.color.Color;
@@ -52,7 +52,7 @@ public class PieDataset {
 	/**
 	 * @see #setBorderWidth(List)
 	 */
-	private List<BigDecimal> borderWidth;
+	private List<Integer> borderWidth;
 
 	/**
 	 * @see #setHoverBackgroundColor(List)
@@ -67,7 +67,7 @@ public class PieDataset {
 	/**
 	 * @see #setHoverBorderWidth(List)
 	 */
-	private List<BigDecimal> hoverBorderWidth;
+	private List<Integer> hoverBorderWidth;
 
 	/**
 	 * @see #setData(List)
@@ -128,14 +128,14 @@ public class PieDataset {
 	/**
 	 * @see #setBorderWidth(List)
 	 */
-	public List<BigDecimal> getBorderWidth() {
+	public List<Integer> getBorderWidth() {
 	    return this.borderWidth;
 	}
 
 	/**
 	 * Border width of arcs in pixels
 	 */
-	public void setBorderWidth(List<BigDecimal> borderWidth) {
+	public void setBorderWidth(List<Integer> borderWidth) {
 	    this.borderWidth = borderWidth;
 	}
 
@@ -170,14 +170,14 @@ public class PieDataset {
 	/**
 	 * @see #setHoverBorderWidth(List)
 	 */
-	public List<BigDecimal> getHoverBorderWidth() {
+	public List<Integer> getHoverBorderWidth() {
 	    return this.hoverBorderWidth;
 	}
 
 	/**
 	 * Border width of arc when hovered
 	 */
-	public void setHoverBorderWidth(List<BigDecimal> hoverBorderWidth) {
+	public void setHoverBorderWidth(List<Integer> hoverBorderWidth) {
 	    this.hoverBorderWidth = hoverBorderWidth;
 	}
